Moved EventDTO to right package
