+++ src/main/java/ch/wisv/areafiftylan/service/EventService.java
@@ -1,7 +1,7 @@
 package ch.wisv.areafiftylan.service;
 
+import ch.wisv.areafiftylan.dto.EventDTO;
 import ch.wisv.areafiftylan.model.Event;
-import ch.wisv.areafiftylan.model.EventDTO;
 import ch.wisv.areafiftylan.model.Team;
 
 import java.util.Collection;
